feat(core): shadow merge the plain object prop when it's necessary
When the parent component passes prop to the child component, if the prop is a plain object, and the
child component's default value build function also returns a plain object, we should shadow merge
the default object and the prop object and add watcher.

BREAKING CHANGE: Now when the default value and giving value are both plain object, they will merge
with the first level.

https://github.com/vuejs/vue/issues/6556